fixed api HomeRoutes path
