+++ controllers/index.js
@@ -1,9 +1,9 @@
 const router = require('express').Router();
 
 const apiRoutes = require('./api');
-const homeRoutes = require('./HomeRoutes.js');
+const homeRoutes = require('./HomeRoutes');
 
-router.use('/', homeRoutes);
-router.use('/api', apiRoutes);
+router.use('/', HomeRoutes);
+// router.use('/api', apiRoutes);
 
 module.exports = router;
