--- serverOld.js
@@ -1,45 +1,0 @@
-require("dotenv").config()
-const path = require('path');
-const express = require('express');
-const session = require('express-session');
-const exphbs = require('express-handlebars');
-const SequelizeStore = require('connect-session-sequelize')(session.Store);
-
-
-
-const routes = require('./controllers');
-const helpers = require('./utils/helpers');
-const sequelize = require('./config/connection');
-
-const app = express();
-const PORT = process.env.PORT || 3000;
-
-const sess = {
-  secret: process.env.SECRET,
-  cookie: {},
-  resave: false,
-  saveUninitialized: true,
-  store: new SequelizeStore({
-    db: sequelize,
-  }),
-};
-
-app.use(session(sess));
-
-// Create the Handlebars.js engine object with custom helper functions
-const hbs = exphbs.create({ helpers });
-
-// Inform Express.js which template engine we're using
-app.engine('handlebars', hbs.engine);
-app.set('view engine', 'handlebars');
-
-// Middleware
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-app.use(express.static(path.join(__dirname, 'public')));
-
-app.use(routes);
-
-sequelize.sync({ force: false }).then(() => {
-  app.listen(PORT, () => console.log('Now listening to port: ' + PORT));
-});
