+++ views/partials/coffee.handlebars
